+++ src/test/kotlin/com/zepben/auth/server/JWTAuthenticatorTest.kt
@@ -10,7 +10,9 @@ package com.zepben.auth.server
 
 import com.auth0.jwk.Jwk
 import com.auth0.jwk.JwkException
+import com.auth0.jwk.SigningKeyNotFoundException
 import com.auth0.jwk.UrlJwkProvider
+import com.auth0.jwt.JWT
 import com.auth0.jwt.exceptions.*
 import com.zepben.auth.client.ProviderDetails
 import com.zepben.auth.common.StatusCode
@@ -19,21 +21,28 @@ import com.zepben.testutils.auth.*
 import com.zepben.testutils.exception.ExpectException.Companion.expect
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.mockkStatic
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
 import org.hamcrest.Matchers.instanceOf
 import org.junit.jupiter.api.Test
 
+fun createAuthenticator(aud: String, issuer: String): JWTAuthenticator {
+    return JWTAuthenticator(
+        aud, listOf(TrustedIssuer(issuer, ProviderDetails("dunno", "https://whereileftmy/keys/"))),
+        verifierBuilder = JWTMultiIssuerVerifierBuilder(
+            aud,
+            listOf(TrustedIssuer(issuer, ProviderDetails("dunno", "https://whereileft.my/keys/"))),
+            true,
+            JWKHolder(true) { _ -> MockJwksUrlProvider().all.associateBy { it.id } })
+    )
+}
+
 class JWTAuthenticatorTest {
 
     @Test
     fun testAuth() {
-        var ta = JWTAuthenticator("https://fake-aud/", listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileftmy/keys/"))),
-            verifierBuilder = JWTMultiIssuerVerifierBuilder(
-                "https://fake-aud/",
-                listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileft.my/keys/"))),
-                true,
-                JWKHolder(true) { _ -> MockJwksUrlProvider().all.associateBy { it.id } } ))
+        var ta = createAuthenticator("https://fake-aud/", "https://issuer/")
 
         var authResp = ta.authenticate(TOKEN)
         assertThat(authResp.statusCode, equalTo(StatusCode.OK))
@@ -61,29 +70,46 @@ class JWTAuthenticatorTest {
         assertThat(authResp.statusCode, equalTo(StatusCode.UNAUTHENTICATED))
         assertThat(authResp.cause, instanceOf(TokenExpiredException::class.java))
 
-        ta = JWTAuthenticator("https://wrong-aud/", listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileftmy/keys/"))),
-            verifierBuilder = JWTMultiIssuerVerifierBuilder(
-                "https://wrong-aud/",
-                listOf(TrustedIssuer("https://issuer/", ProviderDetails("dunno", "https://whereileft.my/keys/"))),
-                true,
-                JWKHolder(true) { _ -> MockJwksUrlProvider().all.associateBy { it.id } } ))
+        ta = createAuthenticator("https://wrong-aud/", "https://issuer/")
 
         authResp = ta.authenticate(TOKEN)
         assertThat(authResp.statusCode, equalTo(StatusCode.PERMISSION_DENIED))
         assertThat(authResp.cause, instanceOf(InvalidClaimException::class.java))
         assertThat(authResp.message, equalTo("The Claim 'aud' value doesn't contain the required audience."))
 
-        ta = JWTAuthenticator("https://fake-aud/", listOf(TrustedIssuer("wrong-issuer", ProviderDetails("dunno", "https://whereileftmy/keys/"))),
-            verifierBuilder = JWTMultiIssuerVerifierBuilder(
-                "https://fake-aud/",
-                listOf(TrustedIssuer("wrong-issuer", ProviderDetails("dunno", "https://whereileft.my/keys/"))),
-                true,
-                JWKHolder(true) { _ -> MockJwksUrlProvider().all.associateBy { it.id } } ))
+        ta = createAuthenticator("https://fake-aud/", "wrong-issuer")
 
         authResp = ta.authenticate(TOKEN)
         assertThat(authResp.statusCode, equalTo(StatusCode.PERMISSION_DENIED))
         assertThat(authResp.cause, instanceOf(InvalidClaimException::class.java))
         assertThat(authResp.message, equalTo("Unknown or untrusted issuer: https://issuer/"))
+    }
+
+    @Test
+    fun `authenticator returns unauthenticated on JwkExceptions`() {
+        val ta = createAuthenticator("https://fake-aud/", "https://issuer/")
+
+        mockkStatic(JWT::class) {
+            every { JWT.decode(any()) } throws SigningKeyNotFoundException("some message", null)
+
+            val authResp = ta.authenticate(TOKEN)
+            assertThat(authResp.statusCode, equalTo(StatusCode.UNAUTHENTICATED))
+            assertThat(authResp.cause, instanceOf(SigningKeyNotFoundException::class.java))
+            assertThat(authResp.message, equalTo("some message"))
+        }
+    }
+
+    @Test
+    fun `authenticator throws exception on unhandled error`() {
+        val ta = createAuthenticator("https://fake-aud/", "https://issuer/")
+
+        mockkStatic(JWT::class) {
+            every { JWT.decode(any()) } throws Exception("some message")
+
+            expect {
+                ta.authenticate(TOKEN)
+            }.toThrow<Exception>()
+        }
     }
 
     @Test
@@ -102,4 +128,6 @@ class JWTAuthenticatorTest {
         }.toThrow(JwkException::class.java)
             .withMessage("Unable to find key fakekey in jwk endpoint. Check your JWK URL.")
     }
+
+
 }